requirements.txt
streamlit
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
transformers==4.38.2
sentence-transformers==2.2.2
huggingface_hub==0.19.4
faiss-cpu
pandas
numpy==1.26.4
tqdm